+++ core/node/via_btc_sender/src/lib.rs
@@ -1,6 +1,6 @@
 mod aggregated_operations;
 mod aggregator;
 mod btc_inscription_aggregator;
-mod btc_tx_manager;
+mod btc_inscription_manager;
 mod config;
 mod publish_criterion;
